[rb] Resolve `uri` gem deprecation warning (#14770)
> /home/user/.rbenv/versions/3.3.6/lib/ruby/gems/3.3.0/gems/selenium-webdriver-4.26.0/lib/selenium/webdriver/remote/bridge.rb:679: warning: URI::RFC3986_PARSER.escape is obsoleted. Use URI::RFC2396_PARSER.escape explicitly.

Ruby switches the default parser from RFC2396 to RFC3986. This parser contains
all methods from the old parser but delegates to RFC2396 for a few and warns.
Namely `extract`, `make_regexp`, `escape`, and `unescape`.

selenium currently supports Ruby >= 3.1, so the current old compatibility code is unnecessary.
`RFC2396_PARSER` is a somewhat new addition, so some kind of check is still needed.
It is available in Ruby 3.1 but early patch versions are missing it.